+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # DF Architect
 
+## Release 5.0.0 (2025-03-10)
+- **UPDATE:** Migrated to v12.
+- **UPDATE:** Downgraded TS -> JS (such sad).
+
 ## Release 4.0.0 (2022-10-01)
 - **NEW:** 90% of the core features have been migrated to V10. All but the Canvas Capture and Tile Flattener.
 - **UPDATE:** Made the wall counter hint a little more descriptive of all the different wall types.
 - **UPDATE:** When you enable the option to display walls, lights, or sounds on other layers, their layer icons will slowly flash blue to indicate this.
+- **UPDATE:** Spanish Translations courtesy of [lozalojo](https://github.com/lozalojo).
 - **TEMP:** Temporarily disabled Canvas Capture ("Save Game Board Image").
 - **TEMP:** Temporarily disabled Tile Flattener.